scripts: add tooling to automatically keep code from `all-types` up to date
