+++ .vscode/launch.json
@@ -10,7 +10,8 @@
       "request": "launch",
       "program": "${workspaceFolder}/scripts/gsrc/update-from-decomp.py",
       "console": "integratedTerminal",
-      "cwd": "${workspaceFolder}"
+      "cwd": "${workspaceFolder}",
+      "args": ["--game", "jak2", "--file", "joint-mod-h"]
     },
   ]
 }
