+++ Taskfile.yml
@@ -87,6 +87,9 @@ tasks:
     cmds:
       - rm ./decompiler_out/**/*.asm
       - rm ./decompiler_out/**/*disasm.gc
+  update-gsrc: # TODO - make this use git to update all files changed in _REF
+    cmds:
+      - python ./scripts/gsrc/update-from-decomp.py --game "{{.GAME}}" --file {{.FILE}}
   # TOOLS
   analyze-ee-memory:
     cmds:
@@ -108,11 +111,6 @@ tasks:
         ignore_error: true
       - python ./scripts/update_decomp_reference.py ./failures ./test/decompiler/reference/
       - task: offline-tests
-  # find-label-types:
-  #   cmds:
-  #     - python ./scripts/next-decomp-file.py --files "{{.FILES}}"
-  #     - task: decomp
-  #     - python ./scripts/find-label-types.py --file "{{.FILES}}"
   # check-gsrc-file:
   #   cmds:
   #     - python ./scripts/check-gsrc-file.py --files "{{.FILES}}"
