+++ decompiler/config/jak2/all-types.gc
@@ -7526,6 +7526,7 @@
 (define-extern *display-deci-count* symbol)
 (define-extern *sync-dma* symbol)
 
+;; +++main-h:strip-lines-controls
 (defenum strip-lines-controls
   :type int64
   (none 0)
@@ -7542,6 +7543,7 @@
   (all-edges 255)
   (strips 256)
   (frags 512))
+;; ---main-h:strip-lines-controls
 
 (define-extern *display-strip-lines* strip-lines-controls)
 
@@ -7592,6 +7594,7 @@
 
 (define-extern *bug-report-output-mode* symbol)
 
+;; +++main-h:scene-controls
 (defenum scene-controls
   :type int64
   (all-off 0)
@@ -7603,18 +7606,22 @@
   (actor-marks 32)
   (special-fma-spheres 64)
   (all-on 95))
+;; ---main-h:scene-controls
 
 (define-extern *display-scene-control* scene-controls)
 
+;; +++main-h:bot-marks-controls
 (defenum bot-marks-controls
   :type int64
   (all-off 0)
   (course-spots 1)
   (task-spots 2)
   (all-on 3))
+;; ---main-h:bot-marks-controls
 
 (define-extern *display-bot-marks* bot-marks-controls)
 
+;; +++main-h:race-marks-controls
 (defenum race-marks-controls
   :type int64
   (all-off 0)
@@ -7627,10 +7634,12 @@
   (path6-orange 64)
   (path7-black 128)
   (all-paths-on 255))
+;; ---main-h:race-marks-controls
 
 (define-extern *display-race-marks* race-marks-controls)
 (define-extern *race-record-path* symbol)
 
+;; +++main-h:race-selection
 (defenum race-selection
   :type int64
   (kiera-class3 0)
@@ -7645,23 +7654,27 @@
   (bush-class3-reverse 9)
   (bush-class2-reverse 10)
   (bush-class1-reverse 11))
+;; ---main-h:race-selection
 
 (define-extern *select-race* race-selection)
 (define-extern *select-race-path* int)
 (define-extern *bot-record-path* int)
 
+;; +++main-h:subdivide-setting
 (defenum subdivide-setting
   :type int64
   (textured 0)
   (outline 1)
   (gouraud 2)
   (hack 3))
+;; ---main-h:subdivide-setting
 
 (define-extern *subdivide-draw-mode* subdivide-setting)
 (define-extern *subdivide-scissor-draw-mode* subdivide-setting)
 (define-extern *subdivide-foreground-draw-mode* subdivide-setting)
 (define-extern *subdivide-ocean-draw-mode* subdivide-setting)
 
+;; +++main-h:ocean-height-hack
 (defenum ocean-height-hack
   :type int64
   (zero 1)
@@ -7670,6 +7683,7 @@
   (sewer-hi 4)
   (sewer-med 5)
   (sewer-lo 6))
+;; ---main-h:ocean-height-hack
 
 (define-extern *ocean-height-hack* ocean-height-hack)
 
@@ -14251,6 +14265,7 @@
 
   )
 
+;; +++joint-mod-h:track-mode
 (defenum track-mode
   :bitfield #t
   :type uint16
@@ -14262,6 +14277,7 @@
   (no-rotate      5) ;; 32
   (no-scale       6) ;; 64
   )
+;; ---joint-mod-h:track-mode
 
 (deftype joint-mod (basic)
   ((mode joint-mod-mode)
