+++ goal_src/jak2/engine/anim/joint-mod-h.gc
@@ -11,6 +11,7 @@
 
   )
 
+;; +++track-mode
 (defenum track-mode
   :bitfield #t
   :type uint16
@@ -22,6 +23,7 @@
   (no-rotate      5) ;; 32
   (no-scale       6) ;; 64
   )
+;; ---track-mode
 
 ;; DECOMP BEGINS
 
@@ -78,6 +80,7 @@
     )
   )
 
+
 (deftype try-to-look-at-info (basic)
   ((who  uint64  :offset-assert   8)
    (horz float   :offset-assert  16)
@@ -125,20 +128,6 @@
     )
   )
 
-(defmethod inspect joint-mod-wheel ((obj joint-mod-wheel))
-  (when (not obj)
-    (set! obj obj)
-    (goto cfg-4)
-    )
-  (format #t "[~8x] ~A~%" obj (-> obj type))
-  (format #t "~1Tlast-position: #<vector @ #x~X>~%" (-> obj last-position))
-  (format #t "~1Tangle: ~f~%" (-> obj angle))
-  (format #t "~1Tprocess: ~A~%" (-> obj process))
-  (format #t "~1Twheel-radius: ~f~%" (-> obj wheel-radius))
-  (format #t "~1Twheel-axis: ~D~%" (-> obj wheel-axis))
-  (label cfg-4)
-  obj
-  )
 
 (defun joint-mod-wheel-callback ((arg0 cspace) (arg1 transformq))
   (let ((s4-0 (the-as joint-mod-wheel (-> arg0 param1))))
@@ -196,6 +185,7 @@
     (new (symbol type process-drawable int symbol symbol symbol) _type_ 0)
     )
   )
+
 
 (defun joint-mod-set-local-callback ((arg0 cspace) (arg1 transformq))
   (let ((v1-0 (the-as joint-mod-set-local (-> arg0 param1))))
@@ -259,6 +249,7 @@
     )
   )
 
+
 (defun joint-mod-add-local-callback ((arg0 cspace) (arg1 transformq))
   (let ((s4-0 (the-as joint-mod-add-local (-> arg0 param1))))
     (when (-> s4-0 enable)
@@ -314,6 +305,7 @@
     )
   )
 
+
 (defun joint-mod-set-world-callback ((arg0 cspace) (arg1 transformq))
   (let ((v1-0 (the-as joint-mod-set-world (-> arg0 param1))))
     (if (-> v1-0 enable)
@@ -353,6 +345,7 @@
     (new (symbol type process-drawable int symbol) _type_ 0)
     )
   )
+
 
 (defun joint-mod-blend-local-callback ((arg0 cspace) (arg1 transformq))
   (let ((gp-0 (the-as joint-mod-blend-local (-> arg0 param1))))
@@ -406,6 +399,7 @@
     )
   )
 
+
 (defun joint-mod-spinner-callback ((arg0 cspace) (arg1 transformq))
   (with-pp
     (let ((gp-0 (the-as joint-mod-spinner (-> arg0 param1))))
@@ -455,6 +449,7 @@
     (new (symbol type process-drawable int symbol float) _type_ 0)
     )
   )
+
 
 (defun joint-mod-blend-world-callback ((arg0 cspace) (arg1 transformq))
   (rlet ((vf0 :class vf)
@@ -561,6 +556,7 @@
     )
   )
 
+
 (defun joint-mod-rotate-local-callback ((arg0 cspace) (arg1 transformq))
   (let ((v1-0 (the-as joint-mod-rotate-local (-> arg0 param1))))
     (new 'stack-no-clear 'vector)
@@ -609,3 +605,10 @@
     (joint-mod-ik-method-10 () none 10)
     )
   )
+
+
+0
+
+
+
+
