+++ goal_src/jak2/engine/game/main-h.gc
@@ -5,6 +5,7 @@
 ;; name in dgo: main-h
 ;; dgos: ENGINE, GAME
 
+;; +++strip-lines-controls
 (defenum strip-lines-controls
   :type int64
   (none 0)
@@ -21,7 +22,9 @@
   (all-edges 255)
   (strips 256)
   (frags 512))
+;; ---strip-lines-controls
 
+;; +++scene-controls
 (defenum scene-controls
   :type int64
   (all-off 0)
@@ -33,14 +36,18 @@
   (actor-marks 32)
   (special-fma-spheres 64)
   (all-on 95))
+;; ---scene-controls
 
+;; +++bot-marks-controls
 (defenum bot-marks-controls
   :type int64
   (all-off 0)
   (course-spots 1)
   (task-spots 2)
   (all-on 3))
+;; ---bot-marks-controls
 
+;; +++race-marks-controls
 (defenum race-marks-controls
   :type int64
   (all-off 0)
@@ -53,7 +60,9 @@
   (path6-orange 64)
   (path7-black 128)
   (all-paths-on 255))
+;; ---race-marks-controls
 
+;; +++race-selection
 (defenum race-selection
   :type int64
   (kiera-class3 0)
@@ -68,14 +77,18 @@
   (bush-class3-reverse 9)
   (bush-class2-reverse 10)
   (bush-class1-reverse 11))
+;; ---race-selection
 
+;; +++subdivide-setting
 (defenum subdivide-setting
   :type int64
   (textured 0)
   (outline 1)
   (gouraud 2)
   (hack 3))
+;; ---subdivide-setting
 
+;; +++ocean-height-hack
 (defenum ocean-height-hack
   :type int64
   (zero 1)
@@ -84,6 +97,7 @@
   (sewer-hi 4)
   (sewer-med 5)
   (sewer-lo 6))
+;; ---ocean-height-hack
 
 ;; DECOMP BEGINS
